Move PermissionRequestValidator in the ws package
