+++ server/sonar-server/src/main/java/org/sonar/server/permission/PermissionTemplateUpdater.java
@@ -30,7 +30,7 @@ import org.sonar.server.exceptions.BadRequestException;
 import org.sonar.server.user.UserSession;
 
 import static org.sonar.server.permission.PermissionPrivilegeChecker.checkGlobalAdminUser;
-import static org.sonar.server.permission.PermissionRequestValidator.validateProjectPermission;
+import static org.sonar.server.permission.ws.PermissionRequestValidator.validateProjectPermission;
 
 /**
  * @deprecated since 5.2 can be removed when Ruby doesn't rely on PermissionTemplateService
